removed libopennet as a dependency for libconfig
I incorrectly had libopennet listed as a dependency for this
version of libconfig. libconfig on MacPorts is separated into
two portfiles, an older one called "libconfig", which depends
on libopennet, and a newer one called "libconfig-hr". This particular
implementation is the one shown in the portfile "libconfig-hr" and
thus does not depend on libopennet.