New formula: netsed
NetSED is small and handful utility designed to alter the contents of packets forwarded through your network in real time.

It is useful for network hackers in following applications:

* Black-box protocol auditing - whenever there are two or more propriety boxes communicating over an
undocumented protocol, it is possible to enforce changes in ongoing transmissions. This helps to test
if the tested application is secure.

* Fuzz-alike experiments, integrity tests - whenever you want to test stability of the application
and see how it ensures data integrity.

* Other common applications - fooling other people, content filtering etc., choose whatever you
want to.